rename sensorModalUtils and device files to be consistent
