+++ src/view/App.tsx
@@ -3,7 +3,7 @@
 
 "use strict";
 import * as React from "react";
-import Device from "./container/device/device"
+import Device from "./container/device/Device"
 import "./App.css";
 
 class App extends React.Component {
