+++ src/view/components/toolbar/SensorModalUtils.tsx
