+++ src/view/container/device/Device.tsx
